IOS-0027 Paymnet, #CodingPatterns IOS-0027D Test
