--- MyNBPapp/View/SandBox/Payment/Model/Receipt.swift
@@ -1,16 +1,0 @@
-//
-//  Receipt.swift
-//  MyNBPapp
-//
-//  Created by Sebastian Sciuba on 21/09/2024.
-//
-
-import Foundation
-
-struct Receipt {
-    let amount: Double
-    let date: Date
-    let transactionId: String
-    let note: String
-    
-}
