--- MyNBPapp/View/SandBox/Payment/Service/PaymentError.swift
@@ -1,12 +1,0 @@
-//
-//  PaymentError.swift
-//  MyNBPapp
-//
-//  Created by Sebastian Sciuba on 21/09/2024.
-//
-
-import Foundation
-
-enum PaymentError: Error {
-    case failed
-}
