+++ MyNBPapp/View/SandBox/PaymentTest/Payment/Interface/Payment/PaymentProcessor.swift
@@ -1,0 +1,14 @@
+//
+//  PaymentProcessor.swift
+//  MyNBPapp
+//
+//  Created by Sebastian Sciuba on 20/09/2024.
+//
+
+import Foundation
+
+protocol PaymentProcessor {
+    func processPayment(amount: Double)
+}
+
+
