+++ MyNBPapp/View/SandBox/PaymentTest/Payment/Interface/Receipt/ReceiptGenerator.swift
@@ -1,0 +1,14 @@
+//
+//  ReceiptGenerator.swift
+//  MyNBPapp
+//
+//  Created by Sebastian Sciuba on 20/09/2024.
+//
+
+import Foundation
+
+protocol ReceiptGenerator {
+    func generateReceipt(transactionId: String, amount: Double, date: Date) -> Receipt
+}
+
+
