+++ MyNBPapp/View/SandBox/PaymentTest/Payment/Model/Receipt.swift
@@ -1,0 +1,17 @@
+//
+//  Receipt.swift
+//  MyNBPapp
+//
+//  Created by Sebastian Sciuba on 21/09/2024.
+//
+
+import Foundation
+
+struct Receipt {
+    let amount: Double
+    let date: Date
+    let transactionId: String
+    let note: String
+    
+}
+
