+++ MyNBPapp/View/SandBox/PaymentTest/Payment/PaymentSystem/ApplePay/ApplePayPaymentProcessor.swift
@@ -1,0 +1,17 @@
+//
+//  ApplePayPaymentProcessor.swift
+//  MyNBPapp
+//
+//  Created by Sebastian Sciuba on 20/09/2024.
+//
+
+import Foundation
+
+struct ApplePayPaymentProcessor: PaymentProcessor {
+    
+    func processPayment(amount: Double) {
+        print("Processing ApplePay payment of \(amount) dollars")
+    }
+    
+}
+
