+++ MyNBPapp/View/SandBox/PaymentTest/Payment/Service/PaymentError.swift
@@ -1,0 +1,12 @@
+//
+//  PaymentError.swift
+//  MyNBPapp
+//
+//  Created by Sebastian Sciuba on 21/09/2024.
+//
+
+import Foundation
+
+enum PaymentError: Error {
+    case failed 
+}
