+++ MyNBPapp/View/SandBox/PaymentTest/Payment/SubView/Brige/Colors.swift
@@ -1,0 +1,26 @@
+//
+//  Colors.swift
+//  MyNBPapp
+//
+//  Created by Sebastian Sciuba on 27/09/2024.
+//
+
+import SwiftUI
+
+class RedColor: ShapeColor {
+    func paint() -> Color {
+        .red
+    }
+}
+
+class GreenColor: ShapeColor {
+    func paint() -> Color {
+        .green
+    }
+} 
+
+class BlueColor: ShapeColor {
+    func paint() -> Color {
+        .blue
+    }
+}
